requirements.txt
pylast
spotipy
python-dotenv
